python_template/03_flask_jinja/pyproject.toml
[tool.poetry]
name = "python_03_flask_jinja_venv"
version = "0.1.0"
description = "python项目flask框架"
authors = ["libaojie <libaojie@qq.com>"]

[tool.poetry.dependencies]
python = "^3.7"
cx-freeze = "^6.4.2"

lbj_common = {path = "package/wheel/lbj_common-1.0-py3-none-any.whl"}
lbj_db = {path = "package/wheel/lbj_db-1.0-py3-none-any.whl"}
lbj_flask = {path = "package/wheel/lbj_flask-1.0-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
[[tool.poetry.source]]
name = "douban"
url = "https://pypi.doubanio.com/simple/"
default = true
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"